Avoid erroneous "Post without data"
Apache.fec's current() unconditionally creates a new request object. If
it is called twice for the same request then it this will trigger a
reread of the already read post data which leads to the error
"mode_ferite: Post without data".

Use the singleton design pattern to ensure that only one instance of the
request object is created per request.